%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  getGenericS3.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getGenericS3}
\alias{getGenericS3.default}
\alias{getGenericS3}

\title{Get an S3 generic function}

\description{
 Get an S3 generic function.
}

\usage{\method{getGenericS3}{default}(name, ...)}

\arguments{
  \item{name}{The name of the generic function.}
  \item{...}{Not used.}
}

\seealso{
  \code{\link{setGenericS3}}().
  \code{\link{getMethodS3}}().
  \code{\link{isGenericS3}}().
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}



\keyword{programming}
\keyword{methods}
